+++ store/prefix/store.go
@@ -23,22 +23,22 @@ var (
 func NewStore(parent types.KVStore, prefix []byte) Store {
 	return NewGStore(
 		parent, prefix,
-		func(v []byte) bool { return v == nil },
-		func(v []byte) int { return len(v) },
+		types.BytesIsZero,
+		types.BytesValueLen,
 	)
 }
 
 func NewObjStore(parent types.ObjKVStore, prefix []byte) ObjStore {
 	return NewGStore(
 		parent, prefix,
-		func(v any) bool { return v == nil },
-		func(v any) int { return 1 },
+		types.AnyIsZero,
+		types.AnyValueLen,
 	)
 }
 
-// GStore is similar with cometbft/cometbft/libs/db/prefix_db
-// both gives access only to the limited subset of the store
-// for convinience or safety
+// GStore is similar to cometbft/cometbft/libs/db/prefix_db
+// both give access only to the limited subset of the store
+// for convenience or safety
 type GStore[V any] struct {
 	parent types.GKVStore[V]
 	prefix []byte
@@ -75,51 +75,52 @@ func (s GStore[V]) key(key []byte) (res []byte) {
 	return res
 }
 
-// Implements Store
+// GetStoreType implements Store, returning the parent store's type
 func (s GStore[V]) GetStoreType() types.StoreType {
 	return s.parent.GetStoreType()
 }
 
-// Implements CacheWrap
+// CacheWrap implements CacheWrap, returning a new CacheWrap with the parent store as the underlying store
 func (s GStore[V]) CacheWrap() types.CacheWrap {
 	return cachekv.NewGStore(s, s.isZero, s.valueLen)
 }
 
 // CacheWrapWithTrace implements the KVStore interface.
 func (s GStore[V]) CacheWrapWithTrace(w io.Writer, tc types.TraceContext) types.CacheWrap {
+	// We need to make a type assertion here as the tracekv store requires bytes value types for serialization.
 	if store, ok := any(s).(*GStore[[]byte]); ok {
 		return cachekv.NewGStore(tracekv.NewStore(store, w, tc), store.isZero, store.valueLen)
 	}
 	return s.CacheWrap()
 }
 
-// Implements KVStore
+// Get implements KVStore, calls Get on the parent store with the key prefixed with the prefix
 func (s GStore[V]) Get(key []byte) V {
 	res := s.parent.Get(s.key(key))
 	return res
 }
 
-// Implements KVStore
+// Has implements KVStore, calls Has on the parent store with the key prefixed with the prefix
 func (s GStore[V]) Has(key []byte) bool {
 	return s.parent.Has(s.key(key))
 }
 
-// Implements KVStore
+// Set implements KVStore, calls Set on the parent store with the key prefixed with the prefix
 func (s GStore[V]) Set(key []byte, value V) {
 	types.AssertValidKey(key)
 	types.AssertValidValueGeneric(value, s.isZero, s.valueLen)
 	s.parent.Set(s.key(key), value)
 }
 
-// Implements KVStore
+// Delete implements KVStore, calls Delete on the parent store with the key prefixed with the prefix
 func (s GStore[V]) Delete(key []byte) {
 	s.parent.Delete(s.key(key))
 }
 
-// Implements KVStore
-// Check https://github.com/cometbft/cometbft/blob/master/libs/db/prefix_db.go#L106
+// Iterator implements KVStore
+// Check https://github.com/cometbft/cometbft-db/blob/main/prefixdb_iterator.go#L106
 func (s GStore[V]) Iterator(start, end []byte) types.GIterator[V] {
-	newstart := cloneAppend(s.prefix, start)
+	newStart := cloneAppend(s.prefix, start)
 
 	var newend []byte
 	if end == nil {
@@ -128,13 +129,13 @@ func (s GStore[V]) Iterator(start, end []byte) types.GIterator[V] {
 		newend = cloneAppend(s.prefix, end)
 	}
 
-	iter := s.parent.Iterator(newstart, newend)
+	iter := s.parent.Iterator(newStart, newend)
 
 	return newPrefixIterator(s.prefix, start, end, iter)
 }
 
 // ReverseIterator implements KVStore
-// Check https://github.com/cometbft/cometbft/blob/master/libs/db/prefix_db.go#L129
+// Check https://github.com/cometbft/cometbft-db/blob/main/prefixdb_iterator.go#L129
 func (s GStore[V]) ReverseIterator(start, end []byte) types.GIterator[V] {
 	newstart := cloneAppend(s.prefix, start)
 
@@ -170,17 +171,17 @@ func newPrefixIterator[V any](prefix, start, end []byte, parent types.GIterator[
 	}
 }
 
-// Implements Iterator
+// Domain implements Iterator, returning the start and end keys of the prefixIterator.
 func (pi *prefixIterator[V]) Domain() ([]byte, []byte) {
 	return pi.start, pi.end
 }
 
-// Implements Iterator
+// Valid implements Iterator, checking if the prefixIterator is valid and if the underlying iterator is valid.
 func (pi *prefixIterator[V]) Valid() bool {
 	return pi.valid && pi.iter.Valid()
 }
 
-// Implements Iterator
+// Next implements Iterator, moving the underlying iterator to the next key/value pair that starts with the prefix.
 func (pi *prefixIterator[V]) Next() {
 	if !pi.valid {
 		panic("prefixIterator invalid, cannot call Next()")
@@ -192,7 +193,7 @@ func (pi *prefixIterator[V]) Next() {
 	}
 }
 
-// Implements Iterator
+// Key implements Iterator, returning the stripped prefix key
 func (pi *prefixIterator[V]) Key() (key []byte) {
 	if !pi.valid {
 		panic("prefixIterator invalid, cannot call Key()")
@@ -213,7 +214,7 @@ func (pi *prefixIterator[V]) Value() V {
 	return pi.iter.Value()
 }
 
-// Implements Iterator
+// Close implements Iterator, closing the underlying iterator.
 func (pi *prefixIterator[V]) Close() error {
 	return pi.iter.Close()
 }