+++ store/types/store.go
@@ -346,7 +346,6 @@ type CacheWrap interface {
 	// Write syncs with the underlying store.
 	Write()
 
-	// Discard the write set
 	Discard()
 }
 
@@ -498,7 +497,7 @@ type ObjectStoreKey struct {
 	name string
 }
 
-// Constructs new ObjectStoreKey
+// NewObjectStoreKey constructs new ObjectStoreKey
 // Must return a pointer according to the ocap principle
 func NewObjectStoreKey(name string) *ObjectStoreKey {
 	return &ObjectStoreKey{
@@ -506,12 +505,12 @@ func NewObjectStoreKey(name string) *ObjectStoreKey {
 	}
 }
 
-// Implements StoreKey
+// Name returns the name of the ObjectStoreKey
 func (key *ObjectStoreKey) Name() string {
 	return key.name
 }
 
-// Implements StoreKey
+// String returns a string representation of the ObjectStoreKey
 func (key *ObjectStoreKey) String() string {
 	return fmt.Sprintf("ObjectStoreKey{%p, %s}", key, key.name)
 }